Significant Alterations to Memory Management
First, instead of calcing the map size, we're just going to make the MMAPed area 128kb, which is way more than we actually need (32 channels of 512 pixels is our expected version)

We also added a device tree, which right now needs to be manually installed. Eventually we'll add that in a script, but for now, it's a manual step.

The device tree entry ensures that we reserve 128kb rooted at 0x80000000, which is C31 for the PRU, which is what we were hard coding our address to.

TL;DR: this should solve our memory corruption problems.